qualityACR/test/qualityACR/ReporterTest.java
package qualityACR;

import static org.junit.Assert.assertTrue;

import org.junit.After;
import org.junit.Before;
import org.junit.Test;

import ij.IJ;
import ij.ImageJ;
import ij.ImagePlus;

public class ReporterTest {
	
	@Before
	public void setUp() throws Exception {
		new ImageJ(ImageJ.NORMAL);
	}

	@After
	public void tearDown() throws Exception {
		// new WaitForUserDialog("Do something, then click OK.").show();

	}

	@Test
	public final void testReporter() {

		Reporter.mainReporter();
		// assertTrue(UtilAyv.compareVectors(expected, result, ""));
	}

}
